requirements.txt
dash==2.17.1
dash-bootstrap-components==1.5.0
dash-auth==2.0.0
plotly==5.17.0
pandas==2.1.4
numpy==1.24.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
gunicorn==21.2.0
requests==2.31.0
Scipy==1.11.4
scikit-learn==1.3.2